% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icampcolors-package.R
\docType{package}
\name{icampcolors-package}
\alias{icampcolors}
\alias{icampcolors-package}
\title{icampcolors: iCaMP Laboratory Color Palette}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
