% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icampcolors-package.R
\docType{package}
\name{icampcolors-package}
\alias{icampcolors}
\alias{icampcolors-package}
\title{icampcolors: iCaMP Laboratory Color Palette}
\description{
Color palettes generated for use in BUMC iCaMP Laboratory Figures Constructed using the 'palettes' package and color tutorials developed by Michael McCarthy. This package "provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, color mixing arithmetic, and color interpolation".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcunha2023/icampcolors}
}

}
\author{
\strong{Maintainer}: Jason Cunha \email{jcunha1@bu.edu}

Authors:
\itemize{
  \item Zachary Milstone
}

}
\keyword{internal}
